Add files via upload
Latest update of the RED package improves on the calculation of the surface location for particles to determine which particles have landed on the ellipsoidal body. Previously this calculation was not entirely correct and it led to cases in which all particles immediately landed on the surface. This current version ensures that this does not happen. If this does happen, chances are the simulation is set up with parameters that are either unrealistic, the particles start too close to the surface, or the initial particle velocity is too slow.